Implemented URLreader Interface
